MAINT-restructure-preprocessing-functions
